Shipe Backend + Docs

+++ backend/app/Tenants/Modules/HRM/Controllers/PositionController.php
@@ -18,6 +18,8 @@ class PositionController extends Controller
 
     public function index(Request $request): JsonResponse
     {
+        $this->authorize('viewAny', Position::class);
+
         $paginator = $this->paginateQuery(Position::query()->orderBy('title'), $request);
 
         return $this->paginatedResponse(PositionResource::class, $paginator, $request);
@@ -25,16 +27,22 @@ class PositionController extends Controller
 
     public function store(StorePositionRequest $request): PositionResource
     {
+        $this->authorize('create', Position::class);
+
         return new PositionResource(Position::create($request->validated()));
     }
 
     public function show(Position $position): PositionResource
     {
+        $this->authorize('view', $position);
+
         return new PositionResource($position);
     }
 
     public function update(StorePositionRequest $request, Position $position): PositionResource
     {
+        $this->authorize('update', $position);
+
         $position->update($request->validated());
 
         return new PositionResource($position);
@@ -42,6 +50,8 @@ class PositionController extends Controller
 
     public function destroy(Position $position): JsonResponse
     {
+        $this->authorize('delete', $position);
+
         $position->delete();
 
         return response()->json(['message' => 'Position removed.'], 200);
